+++ source/DSCResources/MSFT_ADDomainController/MSFT_ADDomainController.psm1
@@ -24,6 +24,11 @@ $script:localizedData = Get-LocalizedData -DefaultUICulture 'en-US'
     .PARAMETER SafemodeAdministratorPassword
         Provide a password that will be used to set the DSRM password. This is a PSCredential.
 
+    .PARAMETER UseExistingAccount
+        Specifies whether to use an existing read only domain controller account.
+
+        Not used in Get-TargetResource.
+
     .NOTES
         Used Functions:
             Name                                            | Module
@@ -50,7 +55,11 @@ function Get-TargetResource
 
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
-        $SafemodeAdministratorPassword
+        $SafemodeAdministratorPassword,
+
+        [Parameter()]
+        [System.Boolean]
+        $UseExistingAccount
     )
 
     Assert-Module -ModuleName 'ActiveDirectory'
@@ -83,7 +92,7 @@ function Get-TargetResource
             if ($domainControllerComputerObject.ManagedBy)
             {
                 $domainControllerManagedByObject = $domainControllerComputerObject.ManagedBy | Get-ADObject -Properties objectSid -Credential $Credential
-                
+
                 $delegateAdministratorAccountName = Resolve-SamAccountName -ObjectSid $domainControllerManagedByObject.objectSid
             }
         }
@@ -115,6 +124,7 @@ function Get-TargetResource
             SafemodeAdministratorPassword       = $SafemodeAdministratorPassword
             SiteName                            = $domainControllerObject.Site
             SysvolPath                          = $serviceNETLOGON.SysVol -replace '\\sysvol$', ''
+            UseExistingAccount                  = $UseExistingAccount
         }
     }
     else
@@ -138,6 +148,7 @@ function Get-TargetResource
             SafemodeAdministratorPassword       = $SafemodeAdministratorPassword
             SiteName                            = $null
             SysvolPath                          = $null
+            UseExistingAccount                  = $UseExistingAccount
         }
     }
 
@@ -202,6 +213,10 @@ function Get-TargetResource
         The parameter `InstallDns` is only used during the provisioning of a domain
         controller. The parameter cannot be used to install or uninstall the DNS
         server on an already provisioned domain controller.
+
+    .PARAMETER UseExistingAccount
+        Specifies whether to use an existing read only domain controller account.
+
     .NOTES
         Used Functions:
             Name                                               | Module
@@ -289,7 +304,11 @@ function Set-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $InstallDns
+        $InstallDns,
+
+        [Parameter()]
+        [System.Boolean]
+        $UseExistingAccount
     )
 
     $getTargetResourceParameters = @{
@@ -393,6 +412,11 @@ function Set-TargetResource
         if ($PSBoundParameters.ContainsKey('InstallDns'))
         {
             $installADDSDomainControllerParameters.Add('InstallDns', $InstallDns)
+        }
+
+        if ($PSBoundParameters.ContainsKey('UseExistingAccount'))
+        {
+            $installADDSDomainControllerParameters.Add('UseExistingAccount', $UseExistingAccount)
         }
 
         if (-not [System.String]::IsNullOrWhiteSpace($InstallationMediaPath))
@@ -671,6 +695,11 @@ function Set-TargetResource
 
         Not used in Test-TargetResource.
 
+    .PARAMETER UseExistingAccount
+        Specifies whether to use an existing read only domain controller account.
+
+        Not used in Test-TargetResource.
+
     .NOTES
         Used Functions:
             Name                          | Module
@@ -747,7 +776,11 @@ function Test-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $InstallDns
+        $InstallDns,
+
+        [Parameter()]
+        [System.Boolean]
+        $UseExistingAccount
     )
 
     Write-Verbose -Message ($script:localizedData.TestingConfiguration -f $env:COMPUTERNAME, $DomainName)